% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col_box_key}
\alias{col_box_key}
\title{Create a colour box-plot like colour key}
\usage{
col_box_key(col.pal = RColorBrewer::brewer.pal(9, "YlOrRd"), limits = c(-1,
  1), xpos = 0, width = 1, col.regions = seq(-1, 1, length.out = 10),
  bg = "grey60", border.col = "grey30", at = c(-1, 1),
  add_locator = FALSE, loc_limits = c(-999, 999), loc_col = "grey60",
  loc_width = 1.5, limit_pad = 0.55, at.labels = c(-1, 1), add = FALSE,
  ...)
}
\arguments{
  \item{col.pal}{Vector of colours to use in legend}

  \item{limits}{Numeric vector of length 2 giving the upper
  and lower postion limits to place the legend.}

  \item{xpos}{X position for left limit of legend box.}

  \item{width}{The width of the legend bar. Goes from 0 to
  this value.}

  \item{col.regions}{Y positions to place colour divisions.
  Should be of length \code{length(col.pal) + 1}}

  \item{bg}{Background colour for any region of the
  rectangle without colour.}

  \item{border.col}{Colour of the legend border.}

  \item{at}{Positions of the ticks.}

  \item{add_locator}{Currently ignored.}

  \item{loc_limits}{Currently ignored.}

  \item{loc_col}{Currently ignored.}

  \item{loc_width}{Currently ignored.}

  \item{limit_pad}{Amount to pad y limits from
  \code{limits}}

  \item{at.labels}{Labels to appyl at \code{at} vector.}

  \item{add}{Add to existing plot?}
}
\description{
Create a colour box-plot like colour key
}
\examples{
col_box_key(at = c(-1, 1), limits = c(-1.5, 1.2), limit_pad = 0)
}

