% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{land}
\alias{land}
\title{Global land polygons}
\description{
These are global land polygons. They were created with:
\code{land <-
readShapePoly("../data/110m-land/110m_land.shp");}
\code{land <- ggplot2::fortify(land)}
}
\keyword{data}

