% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prov_SpatialPolygons.R
\docType{data}
\name{prov_SpatialPolygons}
\alias{prov_SpatialPolygons}
\title{Ecoprovince spatial polygon data}
\description{
These have been joined at the province level for outlining
purposes.
}
\keyword{data}

