% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_class_ext.R
\name{range01}
\alias{range01}
\title{Put data on range from 0 to 1 and assign colours}
\usage{
range01(x)
}
\arguments{
  \item{x}{A numeric vector to scaled between 0 and 1.}
}
\description{
Put data on range from 0 to 1 and assign colours
}

