\name{col_box_key}
\alias{col_box_key}
\title{Create a colour box-plot like colour key}
\usage{
col_box_key(col.pal = brewer.pal(9, "YlOrRd"), limits = c(-3, 0),
  width = 1, col.regions = seq(-1, -0.5, length.out = 10), N = 10,
  bg = "grey60", border.col = "grey30", at = c(-3, -1.5, 0),
  add_locator = FALSE, loc_limits = c(-999, 999), loc_col = "grey60",
  loc_width = 1.5, col.border.col = "grey60", limit_pad = 0.55,
  at.labels = c(-999, 999), ...)
}
\description{
Create a colour box-plot like colour key
}

