% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/land.R
\docType{data}
\name{land}
\alias{land}
\title{Global land polygons}
\description{
These are global land polygons. They were created with:
\code{land <-
readShapePoly("../data/110m-land/110m_land.shp");}
\code{land <- ggplot2::fortify(land)}
}
\keyword{data}

