% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_class_ext.R
\name{map_class_ext}
\alias{map_class_ext}
\title{Make class-level extinction probability maps}
\usage{
map_class_ext(er_dat, min_prov_genera = 20,
  col_pal = RColorBrewer::brewer.pal(9, "YlOrRd"), plot_column = "mean.ext",
  plot_order = c("Mammalia", "Elasmobranchii", "Echinoidea", "Gastropoda",
  "Anthozoa", "Bivalvia"), yticks = c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.5,
  1), ytick_labels = NULL, ylabel = "Intrinsic extinction risk",
  col_range = FALSE, exact_limits = NULL, log_yticks = FALSE,
  fixed_range = FALSE, picture_files = NULL,
  silhouette_coords = list(c(-2.2, -0.75, -1.9, -0, 86)),
  max_range_given = NULL, make_plot = TRUE, ...)
}
\arguments{
  \item{er_dat}{A data frame with the data to plot. Should
  have \code{\link[ggplot2]{fortify}} data from a
  \code{\link[maptools]{readShapePoly}} version of the
  EcoRegion data. Also should have columns \code{N.gen},
  \code{mean.ext}. \code{mean.ext} should contain
  *log*-transformed extinction probability.}

  \item{min_prov_genera}{The minimum contemporary genera
  per province *per class* required to plot.}

  \item{col_pal}{The colour palette to plot with. Should be
  approximately a vector of length 9.}

  \item{plot_column}{A character value giving the name of
  the column to plot.}

  \item{plot_order}{A vector of character giving the order
  of the class panels to plot. Should contain
  \code{c("Mammalia", "Elasmobranchii", "Anthozoa",
  "Gastropoda","Bivalvia", "Decapoda", "Echinoidea")}.}

  \item{yticks}{Locations of y-axis labels.}

  \item{ylabel}{Text to label the colour axis with.}

  \item{col_range}{If \code{TRUE} the colour range will be
  set (on a multiplicative scale) to the same for each
  panel. This can help to avoid overemphasizing small
  differences in some panels. If set to \code{FALSE} then
  each panel will have a colour scale that range from the
  minimum to maximum value for a given panel.}

  \item{exact_limits}{Vector of length 2 giving the exact
  colour limits. Ignored if \code{NULL}}

  \item{log_yticks}{Should the log tick values be log
  transformed when placing them?}

  \item{fixed_range}{Should the colour range be fixed
  through all panels?}

  \item{picture_files}{A vector of file paths containing
  silhouettes to be added to the panels. Should be in a
  format for \code{readPicture} in the \pkg{grImport}
  package.}

  \item{silhouette_coords}{A list of numeric vectors of
  length 4 giving the coordinates to pass to
  \code{readPicture} in the \pkg{grImport} package. x1, y1,
  x2, y2. One list element per panel.}

  \item{max_range_given}{A value giving the maximum
  multiplicative range of the colour scale. If left to
  \code{NULL} then this will be calculated internally.}

  \item{make_plot}{Logical: should the maps be made?}

  \item{...}{Anything extra to pass to \code{col_box_key}.}
}
\description{
This function makes a multipanel figure showing extinction
probability (or other measure) by phylogenetic class.
}

