% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map_hotspots}
\alias{map_hotspots}
\title{Make global intrinsic extinction probability hotspot maps}
\usage{
map_hotspots(er_dat, min_prov_genera = 50, hotspot_thresh = 0.8,
  col_pal = RColorBrewer::brewer.pal(9, "YlOrRd"), plot_column = "mean.ext",
  hotspots = TRUE, col_limits = NULL, add_legend = FALSE, at = c(-1, 1),
  at.labels = c(-1, 1))
}
\arguments{
  \item{er_dat}{A data frame with the data to plot. Should
  have \code{\link[ggplot2]{fortify}} data from a
  \code{\link[maptools]{readShapePoly}} version of the
  EcoRegion data. Also should have columns named
  \code{Halpern}, \code{Burrows}, \code{log_OBIS_records},
  \code{N.gen}, \code{mean.ext}. \code{mean.ext} should
  contain *log*-transformed extinction probability. (note
  that you can chose to plot a different column by
  specifying \code{plot_column}}

  \item{min_prov_genera}{The minimum contemporary genera
  per province required to plot.}

  \item{hotspot_thresh}{Percentile threshold to declare a
  contemporary hotspot.}

  \item{col_pal}{The colour palette to plot with. Should be
  approximately a vector of length 9.}

  \item{plot_column}{The name of the column to plot.}

  \item{hotspots}{Show the contemporary hotspots?}

  \item{col_limits}{\code{NULL} to calculate colour limits
  internally. Set to a numeric vector of length 2 to set
  lower and upper limits for the colour scale. All values
  outside this range will be compressed to the min and max
  values.}

  \item{add_legend}{Add a colour legend? You may need to
  increase the right margin in \code{par} to see this.}

  \item{at}{Locations for ticks on the legend.}

  \item{at.labels}{Labels for colour legend ticks.}
}
\description{
This function plots overall intrinsic extinction
probability and can add hotspots of contemporary risk.
}

